refactor: move score input and total into their own component
this will make state management a little tidier
